# api/openapi.yaml
openapi: 3.1.0
info:
  title: Cloud Photo Delivery API
  version: 0.12.0
  description: API-first backend for the Cloud Photo Delivery SaaS.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /events/{eventID}:
    parameters:
      - $ref: "#/components/parameters/EventID"
    get:
      tags: [Events]
      operationId: getEvent
      summary: Get one event
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventEnvelope"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "404":
          description: Event not found (also for other tenants)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    patch:
      tags: [Events]
      operationId: updateEvent
      summary: Update event fields
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventRequest"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventEnvelope"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "402":
          description: Expiry beyond the plan's retention window
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "409":
          description: The event is expired or archived, so its expiry cannot change
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    delete:
      tags: [Events]
      operationId: deleteEvent
      summary: Soft delete an event
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Deleted
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    deviceKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Device API key issued by `POST /devices` (format `cpd_live_<prefix>_<secret>`).
        `Authorization: Device <key>` is also accepted.
  parameters:
    EventID:
      name: eventID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Envelope:
      type: object
      properties:
        data: {}
        error:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/Error"
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    EventStatus:
      type: string
      enum: [upcoming, active, completed, archived, expired]
...
    Event:
      type: object
      required:
        [
          id,
          name,
          slug,
          status,
          eventDate,
          storageBytes,
          photoCount,
          guestCount,
          expiresAt,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        clientName:
          type: string
        clientEmail:
          type: string
        location:
          type: string
        description:
          type: string
        eventDate:
          type: ["string", "null"]
          format: date
        status:
          $ref: "#/components/schemas/EventStatus"
        coverPhotoId:
          type: ["string", "null"]
          format: uuid
        storageBytes:
          type: integer
          format: int64
        photoCount:
          type: integer
          format: int64
        guestCount:
          type: integer
          format: int64
        expiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateEventRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        eventDate:
          type: string
          format: date
          description: Empty string clears the date
        clientName:
          type: string
        clientEmail:
          type: string
          format: email
        location:
          type: string
        description:
          type: string
        expiresAt:
          type: string
          format: date-time
          description: |
            Empty string resets the expiry to the plan's retention window
            (clears it on unlimited plans). Values later than the retention
            window are rejected with `PLAN_LIMIT_REACHED`. Expired and
            archived events reject expiry changes with
            `INVALID_STATUS_TRANSITION`; extend an expired event to
            reactivate it.
...
    EventEnvelope:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Event"
        error:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/Error"
...